feat : 신고 모델 확장
