+++ src/main/java/com/heartz/byeboo/domain/type/EReportTargetType.java
@@ -1,0 +1,15 @@
+package com.heartz.byeboo.domain.type;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum EReportTargetType {
+
+    COMMENT ("댓글"),
+    COMMON_QUEST("공통 퀘스트 게시글");
+
+    private final String label;
+
+}
